add createUser endpoint
