+++ src/main/java/com/patryklikus/publicchat/Launcher.java
@@ -2,7 +2,7 @@
 package com.patryklikus.publicchat;
 
 import com.patryklikus.publicchat.config.BeanInitializer;
-import com.patryklikus.publicchat.https.request.RequestHandlersManager;
+import com.patryklikus.publicchat.https.RequestHandlersManager;
 import com.sun.net.httpserver.HttpServer;
 import java.io.IOException;
 import java.net.InetSocketAddress;
