+++ src/main/java/com/patryklikus/publicchat/controllers/PageController.java
@@ -1,10 +1,11 @@
 /* Copyright Patryk Likus All Rights Reserved. */
 package com.patryklikus.publicchat.controllers;
 
-import com.patryklikus.publicchat.https.request.Request;
-import com.patryklikus.publicchat.https.request.methodMappings.GetMapping;
-import com.patryklikus.publicchat.https.response.Response;
-import com.patryklikus.publicchat.https.response.ResponseStatusCode;
+import static com.patryklikus.publicchat.https.models.ResponseStatusCode.OK;
+
+import com.patryklikus.publicchat.https.annotations.GetMapping;
+import com.patryklikus.publicchat.https.models.Request;
+import com.patryklikus.publicchat.https.models.Response;
 import com.patryklikus.publicchat.services.ReaderService;
 
 public class PageController {
@@ -17,18 +18,18 @@ public class PageController {
     @GetMapping
     public Response home(Request request) {
         String content = readerService.readResource("home.html");
-        return new Response(ResponseStatusCode.OK, content);
+        return new Response(OK, content);
     }
 
     @GetMapping(path = "/login")
     public Response login(Request request) {
         String content = readerService.readResource("login.html");
-        return new Response(ResponseStatusCode.OK, content);
+        return new Response(OK, content);
     }
 
     @GetMapping(path = "/register")
     public Response register(Request request) {
         String content = readerService.readResource("register.html");
-        return new Response(ResponseStatusCode.OK, content);
+        return new Response(OK, content);
     }
 }
