+++ src/main/java/com/patryklikus/publicchat/controllers/UserController.java
@@ -1,11 +1,11 @@
 /* Copyright Patryk Likus All Rights Reserved. */
 package com.patryklikus.publicchat.controllers;
 
-import com.patryklikus.publicchat.https.request.Request;
-import com.patryklikus.publicchat.https.request.methodMappings.PostMapping;
-import com.patryklikus.publicchat.https.request.methodMappings.RequestMapping;
-import com.patryklikus.publicchat.https.response.Response;
-import com.patryklikus.publicchat.https.response.ResponseStatusCode;
+import com.patryklikus.publicchat.https.annotations.PostMapping;
+import com.patryklikus.publicchat.https.annotations.RequestMapping;
+import com.patryklikus.publicchat.https.models.Request;
+import com.patryklikus.publicchat.https.models.Response;
+import com.patryklikus.publicchat.https.models.ResponseStatusCode;
 import com.patryklikus.publicchat.models.User;
 import com.patryklikus.publicchat.models.mappers.ObjectMapper;
 import com.patryklikus.publicchat.services.UserService;
