+++ src/main/java/com/patryklikus/publicchat/https/annotations/DeleteMapping.java
@@ -1,10 +1,10 @@
 /* Copyright Patryk Likus All Rights Reserved. */
-package com.patryklikus.publicchat.https.request.methodMappings;
+package com.patryklikus.publicchat.https.annotations;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 
 @Retention(RetentionPolicy.RUNTIME)
-public @interface RequestMapping {
-    String path() default "";
+public @interface DeleteMapping {
+    String path() default "/";
 }
