+++ src/main/java/com/patryklikus/publicchat/https/annotations/PostMapping.java
@@ -1,5 +1,5 @@
 /* Copyright Patryk Likus All Rights Reserved. */
-package com.patryklikus.publicchat.https.request.methodMappings;
+package com.patryklikus.publicchat.https.annotations;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
