+++ src/main/java/com/patryklikus/publicchat/https/engine/StringResponseSender.java
@@ -1,6 +1,7 @@
 /* Copyright Patryk Likus All Rights Reserved. */
-package com.patryklikus.publicchat.https.response;
+package com.patryklikus.publicchat.https.engine;
 
+import com.patryklikus.publicchat.https.models.Response;
 import com.sun.net.httpserver.HttpExchange;
 import java.io.IOException;
 import java.io.OutputStream;
