+++ src/main/java/com/patryklikus/publicchat/https/models/Request.java
@@ -1,5 +1,5 @@
 /* Copyright Patryk Likus All Rights Reserved. */
-package com.patryklikus.publicchat.https.request;
+package com.patryklikus.publicchat.https.models;
 
 import com.sun.net.httpserver.Headers;
 import com.sun.net.httpserver.HttpExchange;
