+++ src/main/java/com/patryklikus/publicchat/https/models/RequestMethod.java
@@ -1,5 +1,5 @@
 /* Copyright Patryk Likus All Rights Reserved. */
-package com.patryklikus.publicchat.https.request;
+package com.patryklikus.publicchat.https.models;
 
 public enum RequestMethod {
     GET,
