+++ src/main/java/com/patryklikus/publicchat/https/models/Response.java
@@ -1,5 +1,5 @@
 /* Copyright Patryk Likus All Rights Reserved. */
-package com.patryklikus.publicchat.https.response;
+package com.patryklikus.publicchat.https.models;
 
 public record Response(
         ResponseStatusCode code,
