+++ src/main/java/com/patryklikus/publicchat/https/models/ResponseBodyFormat.java
@@ -1,0 +1,6 @@
+/* Copyright Patryk Likus All Rights Reserved. */
+package com.patryklikus.publicchat.https.models;
+
+public enum ResponseBodyFormat {
+    STRING;
+}
