+++ src/main/java/com/patryklikus/publicchat/https/models/ResponseStatusCode.java
@@ -1,5 +1,5 @@
 /* Copyright Patryk Likus All Rights Reserved. */
-package com.patryklikus.publicchat.https.response;
+package com.patryklikus.publicchat.https.models;
 
 /**
  * Enum represents http response status code
