--- src/main/java/com/patryklikus/publicchat/https/response/ResponseBodyFormat.java
@@ -1,9 +1,0 @@
-/* Copyright Patryk Likus All Rights Reserved. */
-package com.patryklikus.publicchat.https.response;
-
-public enum ResponseBodyFormat {
-    STRING,
-    HTML,
-    CSS,
-    JAVASCRIPT;
-}
