+++ src/main/java/com/patryklikus/publicchat/models/PostBuilder.java
@@ -1,3 +1,4 @@
+/* Copyright Patryk Likus All Rights Reserved. */
 package com.patryklikus.publicchat.models;
 
 import java.time.LocalDateTime;
